renaming, because helper delivers now mesos TaskStatus
